rag-service/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.0

# OpenAI
openai==1.58.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-pptx==1.0.2
python-docx==1.1.2

# Video/Audio transcription
openai-whisper==20240930
yt-dlp==2024.12.13

# Async & Background tasks
celery[redis]==5.4.0
redis==5.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
boto3==1.35.86
httpx==0.28.1
tenacity==9.0.0
tiktoken==0.8.0
